장비추가 개선 및 장비타입에 따라 착용
아이템 추가를 쉽게 하기 위해서
Item클래스를 상속받는 Weapon클래스와 Aromor클래스를 따로 파서
실체 클래스가 Weapon 또는 Aromor중 클래스 상속받는 형식으로 수정함

아이템 추가목록
TraineeArmor
BronzeAxe
SpartaArmor

Player클래스에 딕셔너리로
플레이어 장비창 구현
< 나중에 장비창이라는 클래스 따로 만들어서 관리하면 편할것 같음>

+++ SpartaTextRPG/02.Object/Entity.cs
@@ -15,7 +15,7 @@ namespace SpartaTextRPG
     }
     internal abstract class Entity : ICharacter
     {
-        public Entity(string _name, int _level, int _health, int _defence, int _attack) 
+        public Entity(string _name, int _level, int _health, int _defence, float _attack) 
         {
             name = _name;
             level = _level;
@@ -49,8 +49,8 @@ namespace SpartaTextRPG
         private int defence;
         public int Defence { get => defence; set => defence = value; }
 
-        private int attack;
-        public int Attack { get => attack; set => attack = value; }
+        private float attack;
+        public float Attack { get => attack; set => attack = value; }
 
         private EntityType type;
         public EntityType Type { get => type; set => type = value; }
