+++ SpartaTextRPG/Interface/ICharacter.cs
@@ -12,6 +12,6 @@ namespace SpartaTextRPG.Interface
         public string Name { get; set; }
         public int Health { get; set; }
         public int Defence { get; set; }
-        public int Attack { get; set; }
+        public float Attack { get; set; }
     }
 }
